feat: refactor App component to utilize custom hooks for data management, filtering, and pagination
